requirements.txt
django==2.0.2
django-cors-headers==2.1.0
djangorestframework==3.7.7